Convert mkid to ES
