+++ examples/network/pingpong/Makefile
@@ -1,12 +1,12 @@
-MKID=node $(TROUPE)/rt/built/p2p/mkid.js
+MKID=node $(TROUPE)/rt/built/p2p/mkid.mjs
 MKALIASES=node $(TROUPE)/rt/built/p2p/mkaliases.js
 START=$(TROUPE)/network.sh
 
 zero.listener:
-	$(START) zero.trp --id=ids/pingpong-listener.json --rspawn=true --aliases=aliases.json --stdiolev={} # --debug --debugp2p
+	$(START) zero.trp --id=ids/pingpong-listener.json --rspawn=true --aliases=aliases.json --stdiolev={}  --debug --debugp2p
 
 pingpong.dialer:
-	$(START) p2ppingpong.trp --id=ids/pingpong-dialer.json --aliases=aliases.json # --debug --debugp2p
+	$(START) p2ppingpong.trp --id=ids/pingpong-dialer.json --aliases=aliases.json  --debug --debugp2p
 
 create-network-identifiers:
 	mkdir -p ids
