+++ relay/Makefile
@@ -2,4 +2,4 @@ start-relay:
 	./troupe-p2p-relay --key=keys/relay.priv
 
 generate-relay-key:  
-	node $(TROUPE)/rt/built/p2p/mkid.js --privkeyfile=keys/relay.priv --idfile=keys/relay.id --verbose
+	node $(TROUPE)/rt/built/p2p/mkid.mjs --privkeyfile=keys/relay.priv --idfile=keys/relay.id --verbose
